programs/solana-executor/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "solana-executor"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "oracle-shared",
    "redis[hiredis]>=5.0",
    "python-dotenv>=1.0",
    "httpx>=0.27",
    "websockets>=12.0",
    "apscheduler>=3.10",
]

[project.optional-dependencies]
solana = ["solders>=0.21", "solana>=0.34"]

[tool.hatch.build.targets.wheel]
packages = ["solana_executor"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
